requirements.txt
flask==3.0.2
flask-cors==4.0.0
langchain==0.1.11
langchain-anthropic==0.1.4
python-dotenv==1.0.1
requests==2.31.0
dataclasses-json==0.6.4